Introduce Exercises target.
Exercises are tasks that various approaches or conformers can tackle.
The idea is to support naive, hand-written, regex, PEG, consumer, and
arbitrary 3rd party libraries to demonstrate their approaches
and "compete" with each other.

This particular formulation is a straw-person and we may rearchitect
it a bit.
