+++ Sources/Exercises/Participants/PEGParticipant.swift
@@ -1,0 +1,42 @@
+struct PEGParticipant: Participant {
+  static var name: String { "PEG" }
+}
+
+import PEG
+private func graphemeBreakPropertyData(forLine line: String) -> GraphemeBreakEntry? {
+  typealias Pattern = PEG<Character>.Pattern
+
+  /*
+   # This is a comment
+
+   0600..0605    ; Prepend # Cf   [6] ARABIC NUMBER SIGN..ARABIC NUMBER MARK ABOVE
+   06DD          ; Prepend # Cf       ARABIC END OF AYAH
+
+   Decl -> Scalar (".." Scalar)? Space+ ";" Space Property Space "#" .* success
+   Scalar -> \h{4, 6}
+   Space -> \s
+   Property -> \w+
+
+   */
+
+  let scalar = Pattern.repeatRange(.charactetSet(\.isHexDigit), atLeast: 4, atMost: 6)
+  let space = Pattern.charactetSet(\.isWhitespace)
+  let property = Pattern.oneOrMore(.charactetSet(\.isLetter))
+  let entry = Pattern(
+    scalar, .zeroOrOne(Pattern("..", scalar)), .repeat(space, atLeast: 1),
+    ";", space, property, space, "#", .many(.any), .success)
+
+  let program = PEG.Program(start: "Entry", environment: ["Entry": entry])
+
+  let vm = program.compile(for: String.self)
+  let engine = program.transpile(for: String.self)
+  _ = (vm, engine)
+
+  fatalError("Unsupported")
+  //    let resultVM = vm.consume(line)
+  //    let resultTrans = engine.consume(line)
+  //
+  //    precondition(resultVM == resultTrans)
+
+}
+
