+++ Sources/Exercises/Participants/ReferenceParticipant.swift
@@ -1,0 +1,8 @@
+
+struct ReferenceParticipant: Participant {
+  static var name: String { "Reference" }
+
+  static func graphemeBreakProperty() throws -> (String) -> GraphemeBreakEntry? {
+    try! NaiveParticipant.graphemeBreakProperty()
+  }
+}
