+++ Sources/Exercises/Participants/RegexParticipant.swift
@@ -1,0 +1,3 @@
+struct RegexParticipant: Participant {
+  static var name: String { "Regex" }
+}
